hehe

+++ src/main/Main.java
@@ -25,7 +25,7 @@ public class Main {
 		try {
 			Main main = new Main();
 			Game game = new Game();
-			game.setParams(new Constants(main.framerate),(float) resolutionX,(float) resolutionY);
+			game.setParams(new Constants(main.framerate), resolutionX, resolutionY);
 			AppGameContainer app = new AppGameContainer( game );
 			game.app = app;
 			app.setDisplayMode(resolutionX, resolutionY,true);
